+++ src/main/java/org/jabref/model/entry/BibtexString.java
@@ -145,22 +145,14 @@ public class BibtexString implements Cloneable {
      */
     public String getUserComments() {
         if (parsedSerialization != null) {
-
             try {
                 // get the text before the string
                 String prolog = parsedSerialization.substring(0, parsedSerialization.indexOf('@'));
-
-                // delete trailing whitespaces (between string and text)
-                prolog = prolog.replaceFirst("\\s+$", "");
-                // if there is any non whitespace text, write it with proper line separation
-                if (prolog.length() > 0) {
-                    return prolog;
-                }
+                return prolog;
             } catch (StringIndexOutOfBoundsException ignore) {
                 // if this occurs a broken parsed serialization has been set, so just do nothing
             }
         }
-
         return "";
     }