+++ src/main/java/org/jabref/preferences/PreferencesService.java
@@ -238,8 +238,6 @@ public interface PreferencesService {
 
     void storeOpenLastFilesOnStartup(boolean openLastFilesOnStartup);
 
-    NewLineSeparator getNewLineSeparator();
-
     AutoLinkPreferences getAutoLinkPreferences();
 
     boolean shouldAutosave();