+++ src/test/java/org/jabref/logic/bibtex/BibEntryWriterTest.java
@@ -8,6 +8,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 
+import org.jabref.logic.exporter.BibWriter;
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.importer.fileformat.BibtexParser;
@@ -35,20 +36,20 @@ import static org.mockito.Mockito.mock;
 class BibEntryWriterTest {
 
     private static ImportFormatPreferences importFormatPreferences;
-    private BibEntryWriter writer;
+    private StringWriter stringWriter = new StringWriter();
+    private BibWriter bibWriter = new BibWriter(stringWriter, OS.NEWLINE);
+    private BibEntryWriter bibEntryWriter;
     private final FileUpdateMonitor fileMonitor = new DummyFileUpdateMonitor();
 
     @BeforeEach
     void setUpWriter() {
         importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
         FieldWriterPreferences fieldWriterPreferences = mock(FieldWriterPreferences.class, Answers.RETURNS_DEEP_STUBS);
-        writer = new BibEntryWriter(new FieldWriter(fieldWriterPreferences), new BibEntryTypesManager());
+        bibEntryWriter = new BibEntryWriter(new FieldWriter(fieldWriterPreferences), new BibEntryTypesManager());
     }
 
     @Test
     void testSerialization() throws IOException {
-        StringWriter stringWriter = new StringWriter();
-
         BibEntry entry = new BibEntry(StandardEntryType.Article);
         // set a required field
         entry.setField(StandardField.AUTHOR, "Foo Bar");
@@ -57,12 +58,10 @@ class BibEntryWriterTest {
         entry.setField(StandardField.NUMBER, "1");
         entry.setField(StandardField.NOTE, "some note");
 
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
-        String expected = OS.NEWLINE + "@Article{," + OS.NEWLINE +
+        String expected = "@Article{," + OS.NEWLINE +
                 "  author  = {Foo Bar}," + OS.NEWLINE +
                 "  journal = {International Journal of Something}," + OS.NEWLINE +
                 "  note    = {some note}," + OS.NEWLINE +
@@ -70,12 +69,12 @@ class BibEntryWriterTest {
                 "}" + OS.NEWLINE;
         // @formatter:on
 
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
     void writeOtherTypeTest() throws Exception {
-        String expected = OS.NEWLINE + "@Other{test," + OS.NEWLINE +
+        String expected = "@Other{test," + OS.NEWLINE +
                 "  comment = {testentry}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
 
@@ -83,12 +82,8 @@ class BibEntryWriterTest {
         entry.setField(StandardField.COMMENT, "testentry");
         entry.setCitationKey("test");
 
-        // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
-
-        assertEquals(expected, actual);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
@@ -97,11 +92,9 @@ class BibEntryWriterTest {
         LinkedFile file = new LinkedFile("test", Path.of("/home/uers/test.pdf"), "PDF");
         entry.addFile(file);
 
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(OS.NEWLINE +
-                "@Article{,"
+        assertEquals("@Article{,"
                 + OS.NEWLINE
                 + "  file = {test:/home/uers/test.pdf:PDF},"
                 + OS.NEWLINE
@@ -110,8 +103,6 @@ class BibEntryWriterTest {
 
     @Test
     void writeEntryWithOrField() throws Exception {
-        StringWriter stringWriter = new StringWriter();
-
         BibEntry entry = new BibEntry(StandardEntryType.InBook);
         // set an required OR field (author/editor)
         entry.setField(StandardField.EDITOR, "Foo Bar");
@@ -120,12 +111,10 @@ class BibEntryWriterTest {
         entry.setField(StandardField.NUMBER, "1");
         entry.setField(StandardField.NOTE, "some note");
 
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
-        String expected = OS.NEWLINE + "@InBook{," + OS.NEWLINE +
+        String expected = "@InBook{," + OS.NEWLINE +
                 "  editor  = {Foo Bar}," + OS.NEWLINE +
                 "  note    = {some note}," + OS.NEWLINE +
                 "  number  = {1}," + OS.NEWLINE +
@@ -133,13 +122,11 @@ class BibEntryWriterTest {
                 "}" + OS.NEWLINE;
         // @formatter:on
 
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
     void writeEntryWithOrFieldBothFieldsPresent() throws Exception {
-        StringWriter stringWriter = new StringWriter();
-
         BibEntry entry = new BibEntry(StandardEntryType.InBook);
         // set an required OR field with both fields(author/editor)
         entry.setField(StandardField.AUTHOR, "Foo Thor");
@@ -149,12 +136,10 @@ class BibEntryWriterTest {
         entry.setField(StandardField.NUMBER, "1");
         entry.setField(StandardField.NOTE, "some note");
 
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
-        String expected = OS.NEWLINE + "@InBook{," + OS.NEWLINE +
+        String expected = "@InBook{," + OS.NEWLINE +
                 "  author  = {Foo Thor}," + OS.NEWLINE +
                 "  editor  = {Edi Bar}," + OS.NEWLINE +
                 "  note    = {some note}," + OS.NEWLINE +
@@ -163,12 +148,12 @@ class BibEntryWriterTest {
                 "}" + OS.NEWLINE;
         // @formatter:on
 
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
     void writeReallyUnknownTypeTest() throws Exception {
-        String expected = OS.NEWLINE + "@Reallyunknowntype{test," + OS.NEWLINE +
+        String expected = "@Reallyunknowntype{test," + OS.NEWLINE +
                 "  comment = {testentry}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
 
@@ -176,13 +161,9 @@ class BibEntryWriterTest {
         entry.setType(new UnknownEntryType("ReallyUnknownType"));
         entry.setField(StandardField.COMMENT, "testentry");
         entry.setCitationKey("test");
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
-
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
@@ -193,7 +174,7 @@ class BibEntryWriterTest {
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
                 "  Number                   = {1}" + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         // read in bibtex string
@@ -202,11 +183,9 @@ class BibEntryWriterTest {
         BibEntry entry = entries.iterator().next();
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(bibtexEntry, actual);
+        assertEquals(bibtexEntry, stringWriter.toString());
     }
 
     @Test
@@ -217,7 +196,7 @@ class BibEntryWriterTest {
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
                 "  Number                   = {1}" + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         // read in bibtex string
@@ -226,17 +205,63 @@ class BibEntryWriterTest {
         BibEntry entry = entries.iterator().next();
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(bibtexEntry, actual);
+        assertEquals(bibtexEntry.substring(2), stringWriter.toString());
+    }
+
+    @Test
+    void roundTripWithKeepsCRLFLineBreakStyle() throws IOException {
+        // @formatter:off
+        String bibtexEntry = "@Article{test,\r\n" +
+                "  Author                   = {Foo Bar},\r\n" +
+                "  Journal                  = {International Journal of Something},\r\n" +
+                "  Note                     = {some note},\r\n" +
+                "  Number                   = {1}\r\n" +
+                "}\r\n";
+        // @formatter:on
+
+        // read in bibtex string
+        ParserResult result = new BibtexParser(importFormatPreferences, fileMonitor).parse(new StringReader(bibtexEntry));
+        Collection<BibEntry> entries = result.getDatabase().getEntries();
+        BibEntry entry = entries.iterator().next();
+
+        // write out bibtex string
+        // need to reconfigure writer to use "\r\n"
+        bibWriter = new BibWriter(stringWriter, "\r\n");
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
+
+        assertEquals(bibtexEntry, stringWriter.toString());
+    }
+
+    @Test
+    void roundTripWithKeepsLFLineBreakStyle() throws IOException {
+        // @formatter:off
+        String bibtexEntry = "@Article{test,\n" +
+                "  Author                   = {Foo Bar},\n" +
+                "  Journal                  = {International Journal of Something},\n" +
+                "  Note                     = {some note},\n" +
+                "  Number                   = {1}\n" +
+                "}\n";
+        // @formatter:on
+
+        // read in bibtex string
+        ParserResult result = new BibtexParser(importFormatPreferences, fileMonitor).parse(new StringReader(bibtexEntry));
+        Collection<BibEntry> entries = result.getDatabase().getEntries();
+        BibEntry entry = entries.iterator().next();
+
+        // write out bibtex string
+        // need to reconfigure writer to use "\n"
+        bibWriter = new BibWriter(stringWriter, "\n");
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
+
+        assertEquals(bibtexEntry, stringWriter.toString());
     }
 
     @Test
     void roundTripWithModification() throws IOException {
         // @formatter:off
-        String bibtexEntry = OS.NEWLINE + "@Article{test," + OS.NEWLINE +
+        String bibtexEntry = "@Article{test," + OS.NEWLINE +
                 "  Author                   = {Foo Bar}," + OS.NEWLINE +
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
@@ -253,19 +278,17 @@ class BibEntryWriterTest {
         entry.setField(StandardField.AUTHOR, "BlaBla");
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
-        String expected = OS.NEWLINE + "@Article{test," + OS.NEWLINE +
+        String expected = "@Article{test," + OS.NEWLINE +
                 "  author  = {BlaBla}," + OS.NEWLINE +
                 "  journal = {International Journal of Something}," + OS.NEWLINE +
                 "  note    = {some note}," + OS.NEWLINE +
                 "  number  = {1}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
         // @formatter:on
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
@@ -289,12 +312,10 @@ class BibEntryWriterTest {
         entry.setField(StandardField.AUTHOR, "BlaBla");
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
-        String expected = OS.NEWLINE + "@Article{test," + OS.NEWLINE +
+        String expected = "@Article{test," + OS.NEWLINE +
                 "  author       = {BlaBla}," + OS.NEWLINE +
                 "  journal      = {International Journal of Something}," + OS.NEWLINE +
                 "  note         = {some note}," + OS.NEWLINE +
@@ -302,7 +323,7 @@ class BibEntryWriterTest {
                 "  howpublished = {asdf}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
         // @formatter:on
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
@@ -326,12 +347,10 @@ class BibEntryWriterTest {
         entry.setType(StandardEntryType.InProceedings);
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
-        String expectedNewEntry = OS.NEWLINE + "@InProceedings{test," + OS.NEWLINE +
+        String expectedNewEntry = "@InProceedings{test," + OS.NEWLINE +
                 "  author       = {BlaBla}," + OS.NEWLINE +
                 "  note         = {some note}," + OS.NEWLINE +
                 "  number       = {1}," + OS.NEWLINE +
@@ -339,7 +358,7 @@ class BibEntryWriterTest {
                 "  journal      = {International Journal of Something}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
         // @formatter:on
-        assertEquals(expectedNewEntry, actual);
+        assertEquals(expectedNewEntry, stringWriter.toString());
     }
 
     @Test
@@ -359,12 +378,41 @@ class BibEntryWriterTest {
         BibEntry entry = entries.iterator().next();
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
         String actual = stringWriter.toString();
 
-        // Only one appending newline is written by the writer, the rest by FileActions. So, these should be removed here.
-        assertEquals(bibtexEntry.substring(0, bibtexEntry.length() - 1), actual);
+        // Only one appending newline is written by the writer
+        // OS.NEWLINE is used, not the given one
+        assertEquals(bibtexEntry.substring(0, bibtexEntry.length() - 2) + OS.NEWLINE, actual);
+    }
+
+    @Test
+    void roundTripNormalizesNewLines() throws IOException {
+        // @formatter:off
+        String bibtexEntry = "@Article{test,\n" +
+                "  Author                   = {Foo Bar},\r\n" +
+                "  Journal                  = {International Journal of Something},\n" +
+                "  Number                   = {1},\n" +
+                "  Note                     = {some note}\r\n" +
+                "}\n\n";
+        // @formatter:on
+
+        // read in bibtex string
+        ParserResult result = new BibtexParser(importFormatPreferences, fileMonitor).parse(new StringReader(bibtexEntry));
+        Collection<BibEntry> entries = result.getDatabase().getEntries();
+        BibEntry entry = entries.iterator().next();
+
+        // write out bibtex string
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
+        String actual = stringWriter.toString();
+
+        String expected = "@Article{test," + OS.NEWLINE +
+                "  Author                   = {Foo Bar}," + OS.NEWLINE +
+                "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
+                "  Number                   = {1}," + OS.NEWLINE +
+                "  Note                     = {some note}" + OS.NEWLINE +
+                "}" + OS.NEWLINE;
+        assertEquals(expected, actual);
     }
 
     @Test
@@ -375,7 +423,7 @@ class BibEntryWriterTest {
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
                 "  Number                   = {1}" + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         String result = testSingleWrite(bibtexEntry);
@@ -392,10 +440,11 @@ class BibEntryWriterTest {
         BibEntry entry = entries.iterator().next();
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        StringWriter writer = new StringWriter();
+        BibWriter bibWriter = new BibWriter(writer, OS.NEWLINE);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
+        String actual = writer.toString();
         assertEquals(bibtexEntry, actual);
         return actual;
     }
@@ -407,7 +456,7 @@ class BibEntryWriterTest {
                 "  Author                   = {Foo Bar}," + OS.NEWLINE +
                 "  Month                    = mar," + OS.NEWLINE +
                 "  Number                   = {1}" + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         // read in bibtex string
@@ -421,11 +470,9 @@ class BibEntryWriterTest {
         assertEquals("#mar#", entry.getField(StandardField.MONTH).get());
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(bibtexEntry, actual);
+        assertEquals(bibtexEntry, stringWriter.toString());
     }
 
     @Test
@@ -435,11 +482,9 @@ class BibEntryWriterTest {
         // enable writing
         entry.setChanged(true);
 
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(OS.NEWLINE +
-                        "@Misc{," + OS.NEWLINE +
+        assertEquals("@Misc{," + OS.NEWLINE +
                         "  month = apr," + OS.NEWLINE +
                         "}" + OS.NEWLINE,
                 stringWriter.toString());
@@ -452,11 +497,9 @@ class BibEntryWriterTest {
         // enable writing
         entry.setChanged(true);
 
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(OS.NEWLINE +
-                        "@Misc{," + OS.NEWLINE +
+        assertEquals("@Misc{," + OS.NEWLINE +
                         "  month = {apr}," + OS.NEWLINE +
                         "}" + OS.NEWLINE,
                 stringWriter.toString());
@@ -482,9 +525,7 @@ class BibEntryWriterTest {
         entry.setField(StandardField.HOWPUBLISHED, "asdf");
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
         // @formatter:off
         String expected = OS.NEWLINE + "@Article{test," + OS.NEWLINE +
@@ -495,36 +536,30 @@ class BibEntryWriterTest {
                 "  howpublished = {asdf}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
         // @formatter:on
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
     void doNotWriteEmptyFields() throws IOException {
-        StringWriter stringWriter = new StringWriter();
-
         BibEntry entry = new BibEntry(StandardEntryType.Article);
         entry.setField(StandardField.AUTHOR, "  ");
         entry.setField(StandardField.NOTE, "some note");
 
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        String actual = stringWriter.toString();
-
-        String expected = OS.NEWLINE + "@Article{," + OS.NEWLINE +
+        String expected = "@Article{," + OS.NEWLINE +
                 "  note   = {some note}," + OS.NEWLINE +
                 "}" + OS.NEWLINE;
 
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
     void writeThrowsErrorIfFieldContainsUnbalancedBraces() {
-        StringWriter stringWriter = new StringWriter();
-
         BibEntry entry = new BibEntry(StandardEntryType.Article);
         entry.setField(StandardField.NOTE, "some text with unbalanced { braces");
 
-        assertThrows(IOException.class, () -> writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX));
+        assertThrows(IOException.class, () -> bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX));
     }
 
     @Test
@@ -536,7 +571,7 @@ class BibEntryWriterTest {
                 "  Journal                  = {International Journal of Something}," + OS.NEWLINE +
                 "  Note                     = {some note}," + OS.NEWLINE +
                 "  Number                   = {1}" + OS.NEWLINE +
-                "}";
+                "}" + OS.NEWLINE;
         // @formatter:on
 
         // read in bibtex string
@@ -545,11 +580,9 @@ class BibEntryWriterTest {
         BibEntry entry = entries.iterator().next();
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
 
-        assertEquals(bibtexEntry, actual);
+        assertEquals(bibtexEntry, stringWriter.toString());
     }
 
     @Test
@@ -573,11 +606,10 @@ class BibEntryWriterTest {
         entry.setField(StandardField.AUTHOR, "John Doe");
 
         // write out bibtex string
-        StringWriter stringWriter = new StringWriter();
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBTEX);
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBTEX);
+
         // @formatter:off
-        String expected = "% Some random comment that should stay here" + OS.NEWLINE + OS.NEWLINE +
+        String expected = "% Some random comment that should stay here" + OS.NEWLINE +
                 "@Article{test," + OS.NEWLINE +
                 "  author  = {John Doe}," + OS.NEWLINE +
                 "  journal = {International Journal of Something}," + OS.NEWLINE +
@@ -586,13 +618,11 @@ class BibEntryWriterTest {
                 "}" + OS.NEWLINE;
         // @formatter:on
 
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
     void alphabeticSerialization() throws IOException {
-        StringWriter stringWriter = new StringWriter();
-
         BibEntry entry = new BibEntry(StandardEntryType.Article);
         // required fields
         entry.setField(StandardField.AUTHOR, "Foo Bar");
@@ -606,12 +636,10 @@ class BibEntryWriterTest {
         entry.setField(StandardField.YEAR, "2019");
         entry.setField(StandardField.CHAPTER, "chapter");
 
-        writer.write(entry, stringWriter, BibDatabaseMode.BIBLATEX);
-
-        String actual = stringWriter.toString();
+        bibEntryWriter.write(entry, bibWriter, BibDatabaseMode.BIBLATEX);
 
         // @formatter:off
-        String expected = OS.NEWLINE + "@Article{," + OS.NEWLINE +
+        String expected = "@Article{," + OS.NEWLINE +
                 "  author       = {Foo Bar}," + OS.NEWLINE +
                 "  date         = {2019-10-16}," + OS.NEWLINE +
                 "  journaltitle = {International Journal of Something}," + OS.NEWLINE +
@@ -623,7 +651,7 @@ class BibEntryWriterTest {
                 "}" + OS.NEWLINE;
         // @formatter:on
 
-        assertEquals(expected, actual);
+        assertEquals(expected, stringWriter.toString());
     }
 
     @Test
@@ -656,10 +684,10 @@ class BibEntryWriterTest {
         entry2.setField(StandardField.YEAR, "2020");
         entry2.setField(StandardField.CHAPTER, "chapter");
 
-        String output = writer.serializeAll(List.of(entry1, entry2), BibDatabaseMode.BIBLATEX);
+        String output = bibEntryWriter.serializeAll(List.of(entry1, entry2), BibDatabaseMode.BIBLATEX);
 
         // @formatter:off
-        String expected1 = OS.NEWLINE + "@Article{," + OS.NEWLINE +
+        String expected1 = "@Article{," + OS.NEWLINE +
                 "  author       = {Journal Author}," + OS.NEWLINE +
                 "  date         = {2020-11-16}," + OS.NEWLINE +
                 "  journaltitle = {Journal of Words}," + OS.NEWLINE +
@@ -672,7 +700,7 @@ class BibEntryWriterTest {
         // @formatter:on
 
         // @formatter:off
-        String expected2 = OS.NEWLINE + "@Book{," + OS.NEWLINE +
+        String expected2 = "@Book{," + OS.NEWLINE +
                 "  author    = {John Book}," + OS.NEWLINE +
                 "  date      = {2017-12-20}," + OS.NEWLINE +
                 "  title     = {Entry Title}," + OS.NEWLINE +
@@ -684,7 +712,7 @@ class BibEntryWriterTest {
                 "}" + OS.NEWLINE;
         // @formatter:on
 
-        assertEquals(expected1 + expected2, output);
+        assertEquals(expected1 + OS.NEWLINE + expected2, output);
 
     }
 }